Extract app schedule ownership
